Relicense from GPLv3 to AGPLv3.
Since game streaming is so common these days, I think it's worth
addressing the loophole in the GPLv3 that it can exploit:

I'm Not A Lawyer(TM) and all, but my understanding is that a game
streaming service could use a GPLv3 emulator like this one and not
need to disclose its source code because technically the binary never
reaches the user's computer. Rather, only the emulator's output does.

The AGPLv3 specifically prevents this, while still being compatible
with the GPLv3.