Use nearest-neighbour sampling for the mixer.
Nearest and linear filtering are both noisy, but since we are
upsampling to and from frequencies above the limit of human hearing,
this should not be a problem.

Okay, I lied: the PCM chip outputs 32kHz, which is below the human
hearing limit, but I can make the emulator upsample to double that
limit to bypass this issue.